Add "home" icon, rm home link, add projects link
Removes old "home" link in navbar, and instead uses the navbar header,
"JM JANZEN" as the proper way to get back to the homepage. To make this
even clearer, adds a little house picture on the same line.

While here, add a blank "projects" link in navbar, to later be filled
out.

This change has not been tested in all std environments.